Update README.md
# LastCheckpoint

I'm grateful to God to have scaled through this Software Developers Bootcamp. It was a long and challenging but I'm grateful that I aced it by 60%. Not tally ready for the job market yet, but I'm glad that I'm no longer the same person I was six months ago, The knowledge I gained here will keep helping me. Thank you